The current codebase will create an alias for a document type ",TEST" as "TEST" and will then let you create ".TEST" with an alias of "TEST" meaning you are left with two document types of the same type. resulting in the following xml
I have fixed this by amending the validation function.
I also spotted the same bug when copying a document type. I fixed this one by adding the validation controls to it's page and adding it's corresponding validation methods. I also wrapped a Page.IsValid to the function that copies the document type.
I also spotted that the copied document type's name is not change so I fixed this too.

PLEASE NOTE that most of the line changes shown are due to visual studio formatting the page. (tabs instead of spaces etc.)