Fix for U4-5397
Since the default header in editing section have been given styling to
look more as an editing field, it seems to have affected the sidebar
headers too (h1), because these also is located inside parent elements
with a umb-panel-header class.

I also added a tiny, tiny change to U4-5313 as the bottom bar seems to
overlap the border. The breadcrumb at bottom have a height of 30 px in
umb-panel-footer-nav and a top border on 1 px. So it seems right with
the change from 30px to 31px from bottom.