Partial Views and Partial View Macro's should allow for folder hierarchy for organization.  This changes makes them work much like the Scripts and css folders.
I didn't add a Type drop down like the scripts screen has.  This is unnecessary and cumbersome.  Instead it works like other screens and the user can generate a hierarchy in their files by adding /'s in the names.  This is how it works in Stylesheets works and its a lot cleaner.  I added a note of this to the creation dialogs as well.

For example create a view file with the name myFolder/myView and the system will take care of folder creation automatically.  There is also a delete option.

Other errors fixed:
 Selecting a folder in the scripts lists was generating an error.
 Stylesheets would not create a new folder(but they wold make nested children in existing folders.
 Partial and Macro Views would not use the proper elpers for generating syncTree code.