Fix for U4-5236
Changed the color of the small icon next to the tree icon, when the node
is selected. I have created a new variable for a lighter blue color. It
just need to have a bigger contrast to the color @blue: #2e8aea.
And the current colors for @turquoise (listview) and @green (has
unpublished version) just haven't a big enough contrast to the blue
color. Maybe also for @red too (but more clear). I just choose the
lightblue to make it consistency and a little difference to the white
tree icon, but #fff might also work or a very dark color.