U4-4418 - fix issue with tab sorting
Fixed issue with tab sorting, where you are able to drag the rows above
the table header. Ensured that <thead> / <th> is added, so the class
"ui-sortable" is added to <tbody> and only <tbody> will be the
drag-area.
I also added styling for delete button so it's more consistent with
other "delete"-buttons, e.g. for in settings for dropdown lists,
checkbox lists and radionbutton lists. Furthermore the delete-button is
moved inside the ItemTemplate, so the classes easily could be added to
the button and also make the button stay close to the fields.