Add the ability to limit the number of items that can be inserted into a row cell in the grid editor.
Adds the Maximum items field in the row configuration, if set to nothing or -1 will allow the unlimited items
The preview in the prevalue editor now shows a small number in the grid cell preview to show how many items the cell has been set to.
The grid property editor view has been updated to show and hide the (+) button depending whether the cell has passed the limit or not.

Dictionary keys have been added to language files that have grid items in them. da, en, en_us, es, nl