Introduced alias for row configurations
Document types, property editors and a lot of other types in Umbraco both have a name and an alias, where the name is only something visual, and the alias is something that can be used in the code. So why not use this for row configurations in the grid as well?

I think an alias for row configurations really makes sense, since there might be a reason for change the name of the row, whereas the alias could stay the same.

The label probably needs some localization, since I'm not sure how to do that.