Retina image of avatar
Added an array of 3 image sizes. The original one (30x30px), one for
screens with DPR (device-pixel-ratio) 2 (60x60px) and one for screens
with DPR 3 (90x90px).

Both for the Umbraco icon and gravatar image.

I have kept the size at 30x30px (the two other should just be 2 and 3
times bigger for retina screens).

Actually the umbraco logo.png is 32x32px ... so I am not sure if it is
better we use 32, 64 and 96px icons? In that case just make sure to
update the less file, and ensure both Umbraco logo and gravatar image
use same sizes.

For now I stick to these changes.