HealthCheck: Two custom checks added for testing purposes
One check wants *THE* color to be green, while the second check wants *THE* color to be blue. Rectify one, and the a new check for the other will fail :D

The two checks are just for testing purposes, so they should obviously be removed before release. But for now they serve at an example that rectifying (executing an action) actually works.