Update markdowneditor.controller.js
This is a important fix for the Markdown property type.  I added code previously to get it to properly record changes made via the button controls.  This was causing every document loaded with a markdown property on it to immedetly be flagged as checked with IsDirty() instead of only when a change was actually made.