U4-8592 Decimal Publish Bug: Behaves differently when publishing through Edit Content vs the API
The properties coming from the save event, are sent on to publishing without any potential changes happening upon saving to the database.
One of these changes is the value doesn't actually become a real decimal value before it is persisted so it would get published as the value typed in the editor field instead of what is actually now persisted to the database.