Pass the $element scope to the Markdown library.  The Markdown code tries to load and bind controls by ID.  In some cases the developers my create a setup with repeating or nested datatypes where multiple markdown fields exists with the same allies.
Instead of doing some complex change to ensure that the control ID"s are unique, this change simply passes the current controllers $element scoped to its html to the library.  Then the library uses the jQuery $("#Id", $element) call to load the proper control from within the scope instead of document.getElementById() which is global.