Sort language by name
The issue was that the web API was returning a dictionary that was interpreted as data object by AngularJS, with each key becoming a property of an enormous object, so when binding this object to the select element, it was sorted by property name, thus the key of the array, which was the And being an object it was impossible to sort it again by value.
Only option was to convert it to an array. And since the original list of culture is already sorted correctly, no need to sort it again.