Updated the Imageprocessor cache location
ImageProcessor by a default configuration value stores it's cached data in a folder within /App_Data/cache, I believe it is a DX issue where there is a folder within App_Data named cache, which is used exclusively by image processor. It should be descriptive as to what it is the cache for! (Especially seeing as Umbraco has many caches!)

I can't think of anything which this would negatively effect, upgrades would be unaffected i believe? 

Any feedback?