Setting hide-alias to false still hides the alias
When using umb-editor-header setting hide-alias to false still hides the alias.

I would expect that it would be visible. The only thing that works to make it visible to is removing the hide-alias attribute.

This PR fixes this. So now you can hide the alias based on property of your scope.

Dave