Fix unable to manually set member password via backoffice
When the site is configured with `allowManuallyChangingPassword` as
"true", changing a members password via the backoffice would display a
validation error stating the old password is required.

To solve the issue, when the "Change Password" button is clicked, the
`reset` value is set to "false". The server side logic now assumes if
the `reset` value is present (true/false) then the backoffice users is
changing the password.

Fixes #4144.