Improve error messages
When debugging deployments I ran into this error message, and I had to attach the Umbraco Core code to figure out which node had an issue being saved. Would be awesome to have more info on which specific content item it is that is failing for when you are looping through this call.