Update CONTRIBUTING.md
I know it's cheesy to update a .md for hacktoberfest but I have been dying to put some things in here and to switch some things around. Perhaps you'll disagree on the 'easiest way to build the source code' but I wanted to invert the two because I find clicking lots easier.