Suggest we try this new SystemSemaphoreReleaser class for releasing the system semaphore.
The existing class seems to have all sorts of nonsense going on with unmanaged handles and deriving from CriticalFinalizerObject. I wouldn't be surprised if somewhere in there the semaphore is being destroyed prematurely.

If you do a build with this I will check it out on my live site with the problem.