Make other entities support updateDate
This feels like the wrong thing to do, and instead should probably be creating a MediaBasic for the UpdateDate (or using MediaSlim) - and then implementing the MediaPicker/entityresource/entityservice etc or even the mediaresource/mediaservice - instead of making all entities that inherit entitybasic support updatedate! But I'm adding it here so it's clear what the impact of this change is, should someone review it.