ModelsBuilder enabled flag not respected
When MB is disabled we were still rebuilding all of nucache models when schema changes are made which is a requirement of PureLive models but if MB is disabled then PureLive models don't matter. This was causing unnecessary performance and db overhead when modifying document types when mb is disabled.