Broadcast 'formSubmitting' event when Macro Parameter Overlay is submitted
Some property editors - eg MultiUrlPicker listen out for the broadcast of the 'formSubmitting' event before setting their value of $scope.model.value... if these editors are used as Macro Parameters they do not save their value. This commit adds the broadcasting of the event when the 'submit' button is called and any parameters have been set.