Ignoring package-lock.json from now on, seems not needed.
Motivation: the only "benefit" I keep seeing is: you will know when dependencies has changed version.
But since we lock our dependencies down quite well in package.json already and we have never even had a problem with a package updating "on it's own" so I don't think it's really needed.
The only thing that updates on it's own is whatever dependencies our dependencies drag along with them, but the main js packages that we ship with are always pinned down to a version we know we want to ship with so that won't be a problem.