Don't log query result to avoid confusion
The query result will always be -1, as the result is the amount of rows affected, but that is not applicable in these create queries. This may cause confusion as devs may wonder why for example 'CreatedIndex' is -1. This also seemed to be the only place where we actually use the result of `database.Execute`.