Fix currentUrls when segments are used
When a segment has been created for some node, the variant count will be
> 1 (invariant + segment). Current code to determine URLs of a node did
not take this into account and assumes > 1 variant means the node is
varied by culture.

So when segments were used without culture variation the "currentUrls"
variable would always end up as []. Umbraco then displays the message
"....." in the Info tab.

We now properly handle the case where we have more than 1 variant but
no language on the current variant. In that case we simply want to show
all urls that are available for the node.