Ensure that upload file in the MediaPicker is returned instead of last file in folder
When you use another sorting for Media items than the default (we sort Media-items automatically by name alphabetically when they are created/uploaded), the onUploadComplete function doesn't return the last the last uploaded file. Instead is returns the last filename. By sorting the files by Id, it ensures that the last added file is returned instead of the last file in the list.