Separate remove button and label for settings and styles in Grid Layout
It's inconvenient that the click area for the remove button laps over the label of settings and styles in the Grid Layout property editor. Simply as the editor may remove the style/setting by accident – and especially since there isn't an undo option.

Known issue: Linebreak of long text labels doesn't appear pretty, but this is cosmetic and can be fixed.