Keep custom claims that are flowed from autolinking
When auto linking with callbacks such as OnExternalLogin, custom claims can be added to the user which are flowed to the auth ticket/identity.
However, when the security stamp validator executes, the identity is re-created manually without any knowledge of those custom claims so
they are lost. This ensures that those custom claims flow through to the re-generated identity during the security stamp validation phase.