Bugfix: Members not locked out after failed login attempts (#9921)
* https://github.com/umbraco/Umbraco-CMS/issues/9861
Added test that shows the error

* https://github.com/umbraco/Umbraco-CMS/issues/9861
Fixed test that shows the error. Caches needs to be active, otherwise it do not show the error

* https://github.com/umbraco/Umbraco-CMS/issues/9861
More fix of test that shows the error. Caches needs to be active, otherwise it do not show the error

* https://github.com/umbraco/Umbraco-CMS/issues/9861
Fix issue by raise event and thereby update caches.

Co-authored-by: Elitsa Marinovska <elm@umbraco.dk>