Removes `inert` attribute when `editorService.closeAll()` is called
Fixes #10708 by removing the `inert` attribute on `#mainWrapper` when `editorService.closeAll()` is called.

This code is necessary because when close all is called, there is no editor passed, and so `removeEditor` is never called, and so the `focusLockService.removeInertAttribute();` method isn't called. I add to the if block that handles the close all code instead.