Update Serilog dependencies
This should at least fix #10793 because of a memory leak fix in Serilog.Sinks.Map 1.0.2. Also updated the rest for good measure.

Note that Serilog.Sinks.File is not upgraded to the latest version because it is a new major (4.1.0 -> 5.0.0). Also note that Serilog.Filters.Expressions is deprecated and should be replaced by Serilog.Expressions. I considered both out of scope for this fix.